Split oversized QA workflow sources
