+++ Sources/Services/MenuBarController.swift
@@ -226,164 +226,4 @@ final class MenuBarController: NSObject, NSPopoverDelegate, NSWindowDelegate {
     }
   }
 
-  // MARK: - Menu Shortcuts (⌘N, ⌘,)
-
-  func installMenuShortcuts() {
-    let mainMenu = NSMenu()
-
-    let appMenu = NSMenu()
-    let appMenuItem = NSMenuItem(title: "RedditReminder", action: nil, keyEquivalent: "")
-    appMenuItem.submenu = appMenu
-    let prefsItem = NSMenuItem(
-      title: Self.settingsMenuItemTitle, action: #selector(handleOpenPreferences),
-      keyEquivalent: ",")
-    prefsItem.target = self
-    appMenu.addItem(prefsItem)
-    appMenu.addItem(NSMenuItem.separator())
-    appMenu.addItem(
-      withTitle: "Quit RedditReminder",
-      action: #selector(NSApplication.terminate(_:)),
-      keyEquivalent: "q"
-    )
-    mainMenu.addItem(appMenuItem)
-
-    let fileMenu = NSMenu(title: "File")
-    let fileMenuItem = NSMenuItem(title: "File", action: nil, keyEquivalent: "")
-    fileMenuItem.submenu = fileMenu
-    let newCaptureItem = NSMenuItem(
-      title: "New Capture", action: #selector(handleNewCapture), keyEquivalent: "n")
-    newCaptureItem.target = self
-    fileMenu.addItem(newCaptureItem)
-    mainMenu.addItem(fileMenuItem)
-
-    #if DEBUG
-      let qaMenu = NSMenu(title: "QA")
-      let qaMenuItem = NSMenuItem(title: "QA", action: nil, keyEquivalent: "")
-      qaMenuItem.submenu = qaMenu
-
-      let copyCaptureItem = NSMenuItem(
-        title: "Copy First Queued Capture",
-        action: #selector(handleQACopyFirstQueuedCapture),
-        keyEquivalent: ""
-      )
-      copyCaptureItem.target = self
-      qaMenu.addItem(copyCaptureItem)
-
-      let copyTitleItem = NSMenuItem(
-        title: "Copy First Queued Capture Title",
-        action: #selector(handleQACopyFirstQueuedCaptureTitle),
-        keyEquivalent: ""
-      )
-      copyTitleItem.target = self
-      qaMenu.addItem(copyTitleItem)
-
-      let copySubmitURLItem = NSMenuItem(
-        title: "Copy First Queued Submit URL",
-        action: #selector(handleQACopyFirstQueuedSubmitURL),
-        keyEquivalent: ""
-      )
-      copySubmitURLItem.target = self
-      qaMenu.addItem(copySubmitURLItem)
-
-      let markPostedItem = NSMenuItem(
-        title: "Mark First Queued Capture Posted",
-        action: #selector(handleQAMarkFirstQueuedCapturePosted),
-        keyEquivalent: ""
-      )
-      markPostedItem.target = self
-      qaMenu.addItem(markPostedItem)
-
-      let markPostedWithURLItem = NSMenuItem(
-        title: "Mark First Queued Capture Posted With URL",
-        action: #selector(handleQAMarkFirstQueuedCapturePostedWithURL),
-        keyEquivalent: ""
-      )
-      markPostedWithURLItem.target = self
-      qaMenu.addItem(markPostedWithURLItem)
-
-      let copyPostedSummaryItem = NSMenuItem(
-        title: "Copy First Posted Capture Summary",
-        action: #selector(handleQACopyFirstPostedCaptureSummary),
-        keyEquivalent: ""
-      )
-      copyPostedSummaryItem.target = self
-      qaMenu.addItem(copyPostedSummaryItem)
-
-      let copyPostedURLItem = NSMenuItem(
-        title: "Copy First Posted URL",
-        action: #selector(handleQACopyFirstPostedURL),
-        keyEquivalent: ""
-      )
-      copyPostedURLItem.target = self
-      qaMenu.addItem(copyPostedURLItem)
-
-      qaMenu.addItem(NSMenuItem.separator())
-
-      let createTestCaptureItem = NSMenuItem(
-        title: "Create Test Capture",
-        action: #selector(handleQACreateTestCapture),
-        keyEquivalent: ""
-      )
-      createTestCaptureItem.target = self
-      qaMenu.addItem(createTestCaptureItem)
-
-      let deleteTestCapturesItem = NSMenuItem(
-        title: "Delete Test Captures",
-        action: #selector(handleQADeleteTestCaptures),
-        keyEquivalent: ""
-      )
-      deleteTestCapturesItem.target = self
-      qaMenu.addItem(deleteTestCapturesItem)
-
-      mainMenu.addItem(qaMenuItem)
-    #endif
-
-    NSApp.mainMenu = mainMenu
-  }
-
-  @objc private func handleNewCapture() {
-    onNewCapture?()
-  }
-
-  @objc private func handleOpenPreferences() {
-    onOpenPreferences?()
-  }
-
-  #if DEBUG
-    @objc private func handleQACopyFirstQueuedCapture() {
-      onQACopyFirstQueuedCapture?()
-    }
-
-    @objc private func handleQACopyFirstQueuedSubmitURL() {
-      onQACopyFirstQueuedSubmitURL?()
-    }
-
-    @objc private func handleQAMarkFirstQueuedCapturePosted() {
-      onQAMarkFirstQueuedCapturePosted?()
-    }
-
-    @objc private func handleQAMarkFirstQueuedCapturePostedWithURL() {
-      onQAMarkFirstQueuedCapturePostedWithURL?()
-    }
-
-    @objc private func handleQACreateTestCapture() {
-      onQACreateTestCapture?()
-    }
-
-    @objc private func handleQADeleteTestCaptures() {
-      onQADeleteTestCaptures?()
-    }
-
-    @objc private func handleQACopyFirstQueuedCaptureTitle() {
-      onQACopyFirstQueuedCaptureTitle?()
-    }
-
-    @objc private func handleQACopyFirstPostedCaptureSummary() {
-      onQACopyFirstPostedCaptureSummary?()
-    }
-
-    @objc private func handleQACopyFirstPostedURL() {
-      onQACopyFirstPostedURL?()
-    }
-  #endif
 }
